Remove dates from post URLs
I started naming post files with the date just so that they would sort
alphabetically. But I don't like the clunky URLs this generates, so I've
removed the date from the filename of each post, but added in an alias
for each existing post so that the old URLs will still work.